Add schema_helpers.rs
