+++ crates/fiber-lib/src/rpc/context.rs
@@ -1,3 +1,5 @@
+use crate::rpc::schema_as_string;
+
 use fiber_types::NodeId;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
@@ -7,6 +9,6 @@ use serde_with::serde_as;
 #[derive(Serialize, Deserialize, Debug, Clone, JsonSchema)]
 pub struct RpcContext {
     /// Node ID, read from user RPC biscuit token
-    #[schemars(schema_with = "crate::rpc::schema_as_string")]
+    #[schemars(schema_with = "schema_as_string")]
     pub node_id: NodeId,
 }
