+++ crates/fiber-lib/src/rpc/info.rs
@@ -1,3 +1,5 @@
+use crate::rpc::{schema_as_string_array, schema_as_uint_hex};
+
 use super::graph::UdtCfgInfos;
 use crate::ckb::CkbConfig;
 use crate::fiber::{NetworkActorCommand, NetworkActorMessage};
@@ -34,7 +36,7 @@ pub struct NodeInfoResult {
     pub node_name: Option<String>,
 
     /// A list of multi-addresses associated with the node.
-    #[schemars(schema_with = "crate::rpc::schema_as_string_array")]
+    #[schemars(schema_with = "schema_as_string_array")]
     pub addresses: Vec<MultiAddr>,
 
     /// The hash of the blockchain that the node is connected to.
@@ -42,12 +44,12 @@ pub struct NodeInfoResult {
 
     /// The minimum CKB funding amount for automatically accepting open channel requests, serialized as a hexadecimal string.
     #[serde_as(as = "U64Hex")]
-    #[schemars(schema_with = "crate::rpc::schema_as_uint_hex")]
+    #[schemars(schema_with = "schema_as_uint_hex")]
     pub open_channel_auto_accept_min_ckb_funding_amount: u64,
 
     /// The CKB funding amount for automatically accepting channel requests, serialized as a hexadecimal string.
     #[serde_as(as = "U64Hex")]
-    #[schemars(schema_with = "crate::rpc::schema_as_uint_hex")]
+    #[schemars(schema_with = "schema_as_uint_hex")]
     pub auto_accept_channel_ckb_funding_amount: u64,
 
     /// The default funding lock script for the node.
@@ -55,32 +57,32 @@ pub struct NodeInfoResult {
 
     /// The locktime expiry delta for Time-Locked Contracts (TLC), serialized as a hexadecimal string.
     #[serde_as(as = "U64Hex")]
-    #[schemars(schema_with = "crate::rpc::schema_as_uint_hex")]
+    #[schemars(schema_with = "schema_as_uint_hex")]
     pub tlc_expiry_delta: u64,
 
     /// The minimum value for Time-Locked Contracts (TLC) we can send, serialized as a hexadecimal string.
     #[serde_as(as = "U128Hex")]
-    #[schemars(schema_with = "crate::rpc::schema_as_uint_hex")]
+    #[schemars(schema_with = "schema_as_uint_hex")]
     pub tlc_min_value: u128,
 
     /// The fee (to forward payments) proportional to the value of Time-Locked Contracts (TLC), expressed in millionths and serialized as a hexadecimal string.
     #[serde_as(as = "U128Hex")]
-    #[schemars(schema_with = "crate::rpc::schema_as_uint_hex")]
+    #[schemars(schema_with = "schema_as_uint_hex")]
     pub tlc_fee_proportional_millionths: u128,
 
     /// The number of channels associated with the node, serialized as a hexadecimal string.
     #[serde_as(as = "U32Hex")]
-    #[schemars(schema_with = "crate::rpc::schema_as_uint_hex")]
+    #[schemars(schema_with = "schema_as_uint_hex")]
     pub channel_count: u32,
 
     /// The number of pending channels associated with the node, serialized as a hexadecimal string.
     #[serde_as(as = "U32Hex")]
-    #[schemars(schema_with = "crate::rpc::schema_as_uint_hex")]
+    #[schemars(schema_with = "schema_as_uint_hex")]
     pub pending_channel_count: u32,
 
     /// The number of peers connected to the node, serialized as a hexadecimal string.
     #[serde_as(as = "U32Hex")]
-    #[schemars(schema_with = "crate::rpc::schema_as_uint_hex")]
+    #[schemars(schema_with = "schema_as_uint_hex")]
     pub peers_count: u32,
 
     /// Configuration information for User-Defined Tokens (UDT) associated with the node.
